+++ neoforge-26.2/src/main/java/io/github/imcinq/wateroptimisation/mixin/client/FluidRendererMixin.java
@@ -74,7 +74,7 @@ public abstract class FluidRendererMixin {
 			method = "tesselate",
 			at = @At(
 					value = "INVOKE",
-					target = "Lnet/minecraft/client/renderer/block/FluidRenderer;shouldRenderFace(Lnet/minecraft/world/level/material/FluidState;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/core/Direction;Lnet/minecraft/world/level/material/FluidState;)Z",
+					target = "Lnet/minecraft/client/renderer/block/FluidRenderer;shouldRenderFace(Lnet/minecraft/world/level/material/FluidState;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/core/Direction;Lnet/minecraft/world/level/block/state/BlockState;)Z",
 					ordinal = 0
 			),
 			cancellable = true,